feat: introduce tool selection and circle support with preview

+++ apps/excalidraw-frontend/draw/index.ts
@@ -1,5 +1,6 @@
 import { BACKEND_URL } from "@/config"
 import axios from "axios"
+import { Sheet } from "lucide-react"
 
 type Shape = {
     id: string,
@@ -9,7 +10,7 @@ type Shape = {
     width: number
     height: number
 } | {
-    id:string,
+    id: string,
     type: "circle",
     centerX: number
     centerY: number
@@ -20,13 +21,12 @@ type Operation = { type: "ADD", shape: Shape } | { type: "DELETE", shapeId: stri
 
 const clientId = Math.random().toString()
 
-
-
 export function initDraw(canvas: HTMLCanvasElement, roomId: string, socket: WebSocket) {
     const ctx = canvas.getContext("2d")
 
     let existingShapes: Shape[] = []
     let undoStack: Operation[] = []
+    let selectedTool: "rect" | "circle" = "rect"
 
     if (!ctx) {
         return
@@ -41,21 +41,23 @@ export function initDraw(canvas: HTMLCanvasElement, roomId: string, socket: WebS
             existingShapes.push(op.shape)
         }
 
-        if(op.type === "DELETE"){
+        if (op.type === "DELETE") {
             existingShapes = existingShapes.filter(x => x.id !== op.shapeId)
         }
         redraw()
     }
 
+    // Whenever the user press ctrl+z this function get hits and it will take the last shape id from the stack put in op. with that op it will call the applyoperation and then applyoperation will do their job
+    // After that it will be send by the wesocket to other users
     const undo = () => {
         const op = undoStack.pop();
 
-        if(!op) return;
+        if (!op) return;
 
         applyOperation(op);
 
         socket.send(JSON.stringify({
-            type:"operation",
+            type: "operation",
             op,
             roomId,
             clientId
@@ -91,14 +93,11 @@ export function initDraw(canvas: HTMLCanvasElement, roomId: string, socket: WebS
         if (res.type === "operation") {
             // Self drawn shape will be ignored
             if (res.clientId === clientId) return;
-
             applyOperation(res.op)
         }
     }
 
     socket.addEventListener("message", handleSocketMessage)
-
-    redraw();
 
     const handleMouseDown = (e: MouseEvent) => {
         clicked = true
@@ -117,7 +116,23 @@ export function initDraw(canvas: HTMLCanvasElement, roomId: string, socket: WebS
         clearCanvas(canvas, existingShapes, ctx)
         ctx.strokeStyle = "#ffffff"
         ctx.lineWidth = 2
-        ctx.strokeRect(startX, startY, width, height)
+
+        if (selectedTool === "rect") {
+            console.log("rect is selected")
+            ctx.strokeRect(startX, startY, width, height)
+        }
+
+        if (selectedTool === "circle") {
+            console.log("circle is selected")
+            const radius = Math.sqrt(
+                Math.pow(pos.x - startX, 2) +
+                Math.pow(pos.y - startY, 2)
+            )
+
+            ctx.beginPath()
+            ctx.arc(startX, startY, radius, 0, Math.PI * 2)
+            ctx.stroke()
+        }
     }
 
     // Draws the Shape locally and send that shape to the server for other users
@@ -129,43 +144,60 @@ export function initDraw(canvas: HTMLCanvasElement, roomId: string, socket: WebS
         const width = pos.x - startX
         const height = pos.y - startY
 
-        const shape: Shape = {
-            id: crypto.randomUUID(),
-            type: "rect",
-            x: startX,
-            y: startY,
-            width,
-            height
+        let shape: Shape | null = null
+        if (selectedTool === "rect") {
+            shape = {
+                id: crypto.randomUUID(),
+                type: "rect",
+                x: startX,
+                y: startY,
+                width,
+                height
+            }
+        } else if (selectedTool === "circle") {
+            const radius = Math.sqrt(
+                Math.pow(pos.x - startX, 2) +
+                Math.pow(pos.y - startY, 2)
+            )
+            shape = {
+                id: crypto.randomUUID(),
+                type: "circle",
+                centerX: startX,
+                centerY: startY,
+                radius
+            }
         }
-
-        const op:Operation = {
-            type:"ADD",
+        if (!shape) return
+        const op: Operation = {
+            type: "ADD",
             shape
         }
         applyOperation(op)
         // While we are creating and adding the shape at the same time we will have a record of undoStack in which the shapeId's of all the shapes will we stored
-    
+
         undoStack.push({
-            type:"DELETE",
-            shapeId:shape.id
+            type: "DELETE",
+            shapeId: shape.id
         })
         console.log("SENDING:", { shape, roomId, clientId })
         socket.send(JSON.stringify({
-                type: "operation",
-                op,
-                roomId,
-                clientId
-            }),
+            type: "operation",
+            op,
+            roomId,
+            clientId
+        }),
         )
     }
 
-    const handleKeyDown = (e:KeyboardEvent) => {
-        if(e.ctrlKey && e.key === "z"){
+    const handleKeyDown = (e: KeyboardEvent) => {
+        if (e.ctrlKey && e.key === "z") {
             undo()
         }
+        if (e.key === 'r' || e.key === 'R') selectedTool = "rect"
+        if (e.key === 'c' || e.key === 'C') selectedTool = "circle"
     }
-    window.addEventListener("keydown", handleKeyDown)
 
+    window.addEventListener("keydown", handleKeyDown)
     canvas.addEventListener("mousedown", handleMouseDown)
     canvas.addEventListener("mousemove", handleMouseMove)
     canvas.addEventListener("mouseup", handleMouseUp)
@@ -199,10 +231,20 @@ function clearCanvas(
 
     existingShapes.forEach((shape) => {
         if (shape.type === "rect") {
-            ctx.fillStyle = "rgba(255, 255, 255)"
             ctx.strokeRect(shape.x, shape.y, shape.width, shape.height)
         }
 
+        if (shape.type === "circle") {
+            ctx.beginPath()
+            ctx.arc(
+                shape.centerX,
+                shape.centerY,
+                shape.radius,
+                0,
+                Math.PI * 2
+            )
+            ctx.stroke()
+        }
     })
 }
